--- src/PerspectiveCameraWithHelper.jsx
@@ -1,9 +1,0 @@
-import { PerspectiveCamera, Helper } from '@react-three/drei';
-import { CameraHelper } from 'three';
-export default function PerspectiveCameraWithHelper({ ...perspectiveCameraProps }) {
-    return (
-      <PerspectiveCamera {...perspectiveCameraProps} >
-      <Helper type={CameraHelper}  />
-    </PerspectiveCamera>
-      );
-}
